refactor(text): consolidate cleanup owners
