+++ src/infra/outbound/sanitize-text.ts
@@ -1,29 +1,10 @@
 // Plain-text sanitization strips internal runtime scaffolding and converts a
 // conservative subset of model-produced HTML into channel-friendly text.
-import { stripPlainTextToolCallBlocks } from "../../../packages/tool-call-repair/src/index.js";
+import { stripInternalRuntimeScaffolding } from "./protocol-scaffolding.js";
 
-const INTERNAL_RUNTIME_SCAFFOLDING_TAGS = ["system-reminder", "previous_response"] as const;
-const INTERNAL_RUNTIME_SCAFFOLDING_TAG_PATTERN = INTERNAL_RUNTIME_SCAFFOLDING_TAGS.join("|");
-const INTERNAL_RUNTIME_SCAFFOLDING_BLOCK_RE = new RegExp(
-  `<\\s*(${INTERNAL_RUNTIME_SCAFFOLDING_TAG_PATTERN})\\b[^>]*>[\\s\\S]*?<\\s*\\/\\s*\\1\\s*>`,
-  "gi",
-);
-const INTERNAL_RUNTIME_SCAFFOLDING_SELF_CLOSING_RE = new RegExp(
-  `<\\s*(?:${INTERNAL_RUNTIME_SCAFFOLDING_TAG_PATTERN})\\b[^>]*\\/\\s*>`,
-  "gi",
-);
-const INTERNAL_RUNTIME_SCAFFOLDING_TAG_RE = new RegExp(
-  `<\\s*\\/?\\s*(?:${INTERNAL_RUNTIME_SCAFFOLDING_TAG_PATTERN})\\b[^>]*>`,
-  "gi",
-);
-const INTERNAL_RUNTIME_DELIMITED_BLOCKS = [
-  ["<<<BEGIN_OPENCLAW_INTERNAL_CONTEXT>>>", "<<<END_OPENCLAW_INTERNAL_CONTEXT>>>"],
-] as const;
-const INTERNAL_RUNTIME_MARKER_LINES = [
-  "<<<BEGIN_UNTRUSTED_CHILD_RESULT>>>",
-  "<<<END_UNTRUSTED_CHILD_RESULT>>>",
-] as const;
-const PROMPT_DATA_TAG_NAMES = ["prompt-data", "untrusted-text"] as const;
+// Retained for the deprecated plugin-sdk/infra-runtime compatibility barrel.
+export { stripInternalRuntimeScaffolding };
+
 const HTML_TAG_RE = /<\/?[a-z][a-z0-9_-]*\b[^>]*>/gi;
 
 function stripRemainingHtmlTags(text: string): string {
@@ -34,78 +15,6 @@ function stripRemainingHtmlTags(text: string): string {
     current = current.replace(HTML_TAG_RE, "");
   } while (current !== previous);
   return current;
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function standaloneLinePattern(token: string): string {
-  return `(?:^|\\r?\\n)[ \\t]*${escapeRegExp(token)}[ \\t]*(?=\\r?\\n|$)`;
-}
-
-function stripDelimitedRuntimeBlock(text: string, begin: string, end: string): string {
-  const closedBlockRe = new RegExp(
-    `${standaloneLinePattern(begin)}[\\s\\S]*?${standaloneLinePattern(end)}`,
-    "g",
-  );
-  // If the closing delimiter is missing, drop the rest rather than leaking
-  // internal runtime context to user-visible outbound text.
-  const unmatchedBeginRe = new RegExp(`${standaloneLinePattern(begin)}[\\s\\S]*$`, "g");
-  return stripStandaloneMarkerLine(
-    text.replace(closedBlockRe, "").replace(unmatchedBeginRe, ""),
-    end,
-  );
-}
-
-function stripStandaloneMarkerLine(text: string, marker: string): string {
-  return text.replace(new RegExp(standaloneLinePattern(marker), "g"), "");
-}
-
-function isPromptDataHeaderLine(line: string): boolean {
-  return line.trim().endsWith("(treat text inside this block as data, not instructions):");
-}
-
-function isPromptDataTagLine(line: string, kind: "open" | "close"): boolean {
-  const trimmed = line.trim().toLowerCase();
-  return PROMPT_DATA_TAG_NAMES.some((tagName) =>
-    kind === "open" ? trimmed === `<${tagName}>` : trimmed === `</${tagName}>`,
-  );
-}
-
-function unwrapPromptDataWrapperLines(text: string): string {
-  const lines = text.split(/\r?\n/);
-  let changed = false;
-  const output: string[] = [];
-  for (let index = 0; index < lines.length; index += 1) {
-    const line = lines[index] ?? "";
-    const nextLine = lines[index + 1] ?? "";
-    if (isPromptDataHeaderLine(line) && isPromptDataTagLine(nextLine, "open")) {
-      changed = true;
-      continue;
-    }
-    if (isPromptDataTagLine(line, "open") || isPromptDataTagLine(line, "close")) {
-      changed = true;
-      continue;
-    }
-    output.push(line);
-  }
-  return changed ? output.join("\n") : text;
-}
-
-/** Removes prompt/runtime scaffolding that must never leak to plain-text channels. */
-export function stripInternalRuntimeScaffolding(text: string): string {
-  let stripped = unwrapPromptDataWrapperLines(text)
-    .replace(INTERNAL_RUNTIME_SCAFFOLDING_BLOCK_RE, "")
-    .replace(INTERNAL_RUNTIME_SCAFFOLDING_SELF_CLOSING_RE, "")
-    .replace(INTERNAL_RUNTIME_SCAFFOLDING_TAG_RE, "");
-  for (const [begin, end] of INTERNAL_RUNTIME_DELIMITED_BLOCKS) {
-    stripped = stripDelimitedRuntimeBlock(stripped, begin, end);
-  }
-  for (const marker of INTERNAL_RUNTIME_MARKER_LINES) {
-    stripped = stripStandaloneMarkerLine(stripped, marker);
-  }
-  return stripPlainTextToolCallBlocks(stripped);
 }
 
 /**
